requirements.txt
# Reproducibility environment: Python 3.10.0 with CUDA 12.8.
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.0+cu128
torchvision==0.22.0+cu128
mmcv==2.2.0
mmengine==0.10.7
mmsegmentation==1.2.2
tensorboard==2.20.0
numpy==2.2.6
opencv-python==4.12.0.88
timm==1.0.24
tqdm==4.67.1
